Remove unused REST API endpoints and code
- Update `config/routes.rb`
- Move `spec/requests/graph/**_spec.rb` to `spec/requests/**_spec.rb`
- Remove unused controllers and their specs
- Remove all active model serializers
