--- app/serializers/access_token_serializer.rb
@@ -1,8 +1,0 @@
-class AccessTokenSerializer < ActiveModel::Serializer
-  embed :ids
-  root false
-
-  attributes :access_token, :expires_in
-
-  has_one :user
-end
