--- app/serializers/score_serializer.rb
@@ -1,7 +1,0 @@
-class ScoreSerializer < ActiveModel::Serializer
-  attributes :id, :points
-
-  embed :ids, include: true
-
-  has_one :user
-end
