--- app/serializers/user_serializer.rb
@@ -1,5 +1,0 @@
-class UserSerializer < ActiveModel::Serializer
-  embed :ids
-
-  attributes :id, :avatar_url, :nickname
-end
