--- app/serializers/weekly_winning_serializer.rb
@@ -1,7 +1,0 @@
-class WeeklyWinningSerializer < ActiveModel::Serializer
-  attributes :id, :start_date, :points
-
-  embed :ids, include: true
-
-  has_one :winner
-end
