+++ spec/requests/users_spec.rb
@@ -1,56 +1,74 @@
 require 'spec_helper'
 
-describe 'Users Requests', type: :request do
-  describe 'GET /users/me' do
-    it 'responds with a json representing the current user' do
-      get me_api_users_path
+describe 'Users graph', type: :request do
+  describe 'query user' do
+    context 'with id "me"' do
+      it 'responds with a json representing the current user' do
+        post api_graph_path(query: <<~QUERY)
+          query user {
+            user(id: "me") {
+              id
+              avatar_url
+              nickname
+            }
+          }
+        QUERY
 
-      expect(last_response.status).to eq(200)
+        expect(JSON.parse(last_response.body)['data']['user']).to eq(
+          'id' => current_user.id,
+          'avatar_url' => current_user.avatar_url,
+          'nickname' => current_user.nickname
+        )
+      end
+    end
 
-      expect(JSON.parse(last_response.body)['user']).to eq(
-        'id' => current_user.id,
-        'avatar_url' => current_user.avatar_url,
-        'nickname' => current_user.nickname
-      )
+    context 'with a regular id' do
+      it 'responds with a json representing the requested user' do
+        user = create(:user)
+
+        post api_graph_path(query: <<~QUERY)
+          query user {
+            user(id: "#{user.id}") {
+              id
+              avatar_url
+              nickname
+            }
+          }
+        QUERY
+
+        expect(JSON.parse(last_response.body)['data']['user']).to eq(
+          'id' => user.id,
+          'nickname' => user.nickname,
+          'avatar_url' => user.avatar_url
+        )
+      end
     end
   end
 
-  describe 'GET /users' do
+  describe 'query users' do
     it 'responds with a json representing the list of registered users' do
       users = create_list(:user, 2)
 
-      get api_users_path
-
-      expect(last_response.status).to eq(200)
-
-      expect(JSON.parse(last_response.body)['users']).to match_array(
-        users.map do |user|
-          {
-            'id' => user.id,
-            'nickname' => user.nickname,
-            'avatar_url' => user.avatar_url
+      post api_graph_path(query: <<~QUERY)
+        query users {
+          users(ids: #{users.map(&:id)}) {
+            id
+            avatar_url
+            nickname
           }
-        end.push(
-          'id' => current_user.id,
-          'nickname' => current_user.nickname,
-          'avatar_url' => current_user.avatar_url
-        )
-      )
-    end
-  end
+        }
+      QUERY
 
-  describe 'GET /user/{id}' do
-    it 'responds with a json representing the requested user' do
-      user = create(:user)
-
-      get api_user_path(user)
-
-      expect(last_response.status).to eq(200)
-
-      expect(JSON.parse(last_response.body)['user']).to eq(
-        'id' => user.id,
-        'nickname' => user.nickname,
-        'avatar_url' => user.avatar_url
+      expect(JSON.parse(last_response.body)['data']['users']).to match_array(
+        [{
+          'id' => users.first.id,
+          'nickname' => users.first.nickname,
+          'avatar_url' => users.first.avatar_url
+        }, {
+          'id' => users.last.id,
+          'nickname' => users.last.nickname,
+          'avatar_url' => users.last.avatar_url
+        }]
       )
     end
   end
