+++ spec/requests/weekly_winnings_spec.rb
@@ -1,31 +1,28 @@
 require 'spec_helper'
 
-describe 'Weekly Winnings Requests', type: :request do
-  describe 'GET /weekly_winnings' do
+describe 'Weekly Winnings graph', type: :request do
+  describe 'query weekly_winnings' do
     it 'responds with a json representing the list of weekly winnings' do
       weekly_winnings = create_list(:weekly_winning, 2)
 
-      get api_weekly_winnings_path
-
-      expect(last_response.status).to eq(200)
-
-      expect(JSON.parse(last_response.body)['weekly_winnings']).to match_array(
-        weekly_winnings.map do |weekly_winning|
-          {
-            'id' => weekly_winning.id,
-            'winner_id' => weekly_winning.winner.id,
-            'start_date' => weekly_winning.start_date.iso8601,
-            'points' => weekly_winning.points
+      post api_graph_path(query: <<~QUERY)
+        query weekly_winnings {
+          weekly_winnings {
+            id
+            points
+            start_date
+            winner_id
           }
-        end
-      )
+        }
+      QUERY
 
-      expect(JSON.parse(last_response.body)['winners']).to match_array(
+      expect(JSON.parse(last_response.body)['data']['weekly_winnings']).to match_array(
         weekly_winnings.map do |weekly_winning|
           {
-            'id' => weekly_winning.winner.id,
-            'avatar_url' => weekly_winning.winner.avatar_url,
-            'nickname' => weekly_winning.winner.nickname
+            'id' => weekly_winning.id.to_s,
+            'start_date' => weekly_winning.start_date.iso8601,
+            'points' => weekly_winning.points,
+            'winner_id' => weekly_winning.winner.id.to_s
           }
         end
       )
